BLEND: Updated temple now using my addon to bake the animation :)
Pendulums hardcoded in a working position, to be dynamically added soon

How to update the animation?
----------------------------
Update the influence values of the 'Trail-Curve' object. This will
affect the animation of the 'Animation_Proxy' object. Now if you are
satisfied with the results, transfer the animation to the 'Proxy' object
with the 'Bake Stalker' operator [1].

Finally, change the action in the actuator of the 'Proxy' object (pick
the latest Stalker.00 created action).

Note: at the moment the animation is created with 10x the played frames
to give precision to the playback script.

[1] -
https://github.com/dfelinto/addon-samples/blob/master/animation_stalker.py